==> Notebooks/1_DataWrangling.ipynb <==
## Importacion de librerias

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import snowflake.connector
from sqlalchemy import create_engine
from sklearn.preprocessing import LabelEncoder
import config 

### Lectura de la base

connection_url = f"snowflake://{config.user}:{config.password}@{config.account}/{config.database}/{config.schema}?warehouse={config.warehouse}"

# Crear el motor de conexión
engine = create_engine(connection_url)

# Verificar la conexión
try:
    with engine.connect() as connection:
        print("Conexión exitosa a Snowflake!")
except Exception as e:
    print(f"Error en la conexión: {e}")

raw = pd.read_sql("SELECT * FROM RAW.BANCO", engine)

raw.info()

Se identifica los valores nulos para decidir cómo deben ser tratados

print('Valores nulos:\n', raw.isnull().sum())

# Se crea una copia del dataframe original para evitar modificarlo directamente
curated = raw.copy()

# Para las variables categoricas, se reemplazan los valores nulos por 'Unknown'
for col in curated.columns:
    if curated[col].dtype == 'object':  
        curated[col] = curated[col].fillna('Unknown') 

# Para las variables numericas, se reemplazan los valores nulos por la media de la 'location' o 'device_used' dependiendo el contexto de lo que se quiere analizar
# Se reemplazan los valores nulos de 'transaction_amount' por la media de la 'location'
# Se reemplazan los valores nulos de 'time_of_transaction' por la media de la 'device_used'
def valores_nulos_num(curated):
    
    curated['transaction_amount'] = curated.groupby('location')['transaction_amount'].transform(lambda x: x.fillna(x.mean()))
    curated['time_of_transaction'] = curated.groupby('device_used')['time_of_transaction'].transform(lambda x: x.fillna(x.mean()))
    
    return curated
    
curated = valores_nulos_num(curated)

Se tiene un dataset curado, sin valores nulos ni duplicados

print('Valores nulos:\n', curated.isnull().sum())

Se exporta a la carpeta curated donde se almacena la base ya limpia y lista para realizar feature engineering

curated.to_csv('D:\Proyecto 2\Data\Clean\curated.csv', index=False)
