==> Notebooks/2_FeatureEngineering.ipynb <==
## Lectura de librerias

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import snowflake.connector
from sqlalchemy import create_engine
from sklearn.preprocessing import LabelEncoder
import config 

Se lee la base previamente limpiada en el proceso anterior

curated = pd.read_csv('D:\Proyecto 2\Data\Clean\curated.csv')
curated.head()

curated.info()

print("Valores unicos para Transaction_ID:", len(curated['transaction_id'].unique()))
print("Valores unicos para User_ID:", len(curated['user_id'].unique()))

Se crea una copia de la base para no modificarla directamente

clean = curated.copy()

# Se decide eliminar estas columna porque no aporta valor al analisis
clean.drop(columns=['transaction_id'], inplace=True)

# Se crea esta nueva variable para saber cuantas transacciones fraudulentas ha tenido el usuario en el pasado y cada cuanto podría volver a tener una
clean['fraud_per_month'] = np.where(
    clean['previous_fraudulent_transactions'] == 0,
    0,
    clean['account_age'] / clean['previous_fraudulent_transactions']
).round(0).astype(int)

clean.head()

#Se crea una copia final del dataset para el modelo de machine learning usando la tecnica de one hot encoding para las variables categoricas y de esta manera poder usar el dataset en el modelado final
ml_dataset = clean.copy()

categorical_features = ['transaction_type', 'device_used', 'location', 'payment_method']

df_train = pd.get_dummies(ml_dataset, columns=categorical_features)

df_train.head()

Se exporta la base final a la carpeta donde se encontrará el dataset a usar para el modelado

df_train.to_csv('D:\Proyecto 2\Data\Processed\ml_dataset.csv', index=False)
