==> Notebooks/3_Modeling.ipynb <==
## Importacion de librerias

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import LabelEncoder

from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, roc_auc_score
from sklearn.linear_model import LogisticRegression
from sklearn.linear_model import LogisticRegression, SGDClassifier
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.tree import DecisionTreeClassifier
from imblearn.over_sampling import SMOTE

Se hace la lectura de la base ya lista para modelar en el proceso anterior

ml_data = pd.read_csv('D:\Proyecto 2\Data\Processed\ml_dataset.csv')
ml_data.head()

ml_data.info()

ml_data['fraudulent'].value_counts()

Debido a la desbalance de clases, se utilizara el algoritmo SMOTE para balancear las clases.

scaler = StandardScaler()
smote = SMOTE(random_state=42)
X_train_resampled, y_train_resampled = smote.fit_resample(ml_data.drop(['fraudulent'], axis=1), ml_data['fraudulent'])
X_scaled = scaler.fit_transform(X_train_resampled)

Se hace el particionamiento de datos

X_entrenamiento_caracteristica, X_prueba_caracteristica, Y_entrenamiento_objetivo, Y_prueba_objetivo = train_test_split(X_scaled, y_train_resampled, random_state=55, stratify=y_train_resampled, test_size=0.2)

### Comenzar modelado con diferentes opciones

reg_log = LogisticRegression(C=1.0, solver="liblinear", max_iter=500, class_weight="balanced", random_state=4525)

reg_log.fit(X_train_resampled, y_train_resampled)

Y_pred = reg_log.predict(X_prueba_caracteristica)


accuracy_rl = accuracy_score(Y_prueba_objetivo, Y_pred)
precision_rl = precision_score(Y_prueba_objetivo, Y_pred, zero_division=0)
recall_rl = recall_score(Y_prueba_objetivo, Y_pred)
f1_rl = f1_score(Y_prueba_objetivo, Y_pred)
auc_roc_rl = roc_auc_score(Y_prueba_objetivo, Y_pred)

# Mostrar las métricas
print(f"Accuracy: {accuracy_rl:.4f}")
print(f"Precisión: {precision_rl:.4f}")
print(f"Recall: {recall_rl:.4f}")
print(f"F1-score: {f1_rl:.4f}")
print(f"AUC-ROC: {auc_roc_rl:.4f}")

resultados_lr = []
resultados_lr.append({
    'Modelo': 'Regresión Logística',
    'Accuracy': accuracy_rl,
    'Precisión': precision_rl,
    'Recall': recall_rl,
    'F1-score': f1_rl,
    'AUC-ROC': auc_roc_rl
})
df_rl = pd.DataFrame(resultados_lr)
df_rl

for profundidad in range(1, 15):
    for split in [2, 5, 10]:
        arbol_clf = DecisionTreeClassifier(max_depth=profundidad, min_samples_split=split, criterion="gini", class_weight="balanced", random_state=42)
        arbol_clf.fit(X_train_resampled, y_train_resampled)
        Y_pred = arbol_clf.predict(X_prueba_caracteristica)

        accuracy_arbol = accuracy_score(Y_prueba_objetivo, Y_pred)
        precision_arbol = precision_score(Y_prueba_objetivo, Y_pred, zero_division=0)
        recall_arbol = recall_score(Y_prueba_objetivo, Y_pred)
        f1_arbol = f1_score(Y_prueba_objetivo, Y_pred)
        auc_roc_arbol = roc_auc_score(Y_prueba_objetivo, Y_pred)

        print(f"Profundidad: {profundidad}, Split: {split}, Accuracy: {accuracy_arbol:.4f}, Precisión: {precision_arbol:.4f}, Recall: {recall_arbol:.4f}, F1-score: {f1_arbol:.4f}, AUC-ROC: {auc_roc_arbol:.4f}")


El mejor resultado para el modelo de árboles de decision es el siguiente: 
Profundidad: 7, Split: 10
Accuracy: 0.9244
Precisión: 0.9611
Recall: 0.8845 
F1-score: 0.9212
AUC-ROC: 0.9244

resultados_decision = []

# Agregar la combinación específica al DataFrame
resultados_decision.append({
    'Modelo': 'Decision Tree',
    "Profundidad": 14,
    "Split": 10,
    "Accuracy": 0.8686,
    "Precision": 0.8374,
    "Recall": 0.9149,
    "F1-Score": 0.8744,
    "AUC-ROC": 0.8686
})

# Convertir la lista de resultados en un DataFrame
df_resultados_decision_tree = pd.DataFrame(resultados_decision)

df_resultados_decision_tree

for estimadores in range(10, 101, 10):
    for profundidad in [5, 10, 15]:
        bosque_clf = RandomForestClassifier(n_estimators=estimadores, max_depth=profundidad, min_samples_split=5, class_weight="balanced_subsample", random_state=42)
        bosque_clf.fit(X_train_resampled, y_train_resampled)
        Y_pred = bosque_clf.predict(X_prueba_caracteristica)

        accuracy_bosque = accuracy_score(Y_prueba_objetivo, Y_pred)
        precision_bosque = precision_score(Y_prueba_objetivo, Y_pred, zero_division=0)
        recall_bosque = recall_score(Y_prueba_objetivo, Y_pred)
        f1_bosque = f1_score(Y_prueba_objetivo, Y_pred)
        auc_roc_bosque = roc_auc_score(Y_prueba_objetivo, Y_pred)

        print(f"Estimadores: {estimadores}, Profundidad: {profundidad}, Accuracy: {accuracy_bosque:.4f}, Precisión: {precision_bosque:.4f}, Recall: {recall_bosque:.4f}, F1-score: {f1_bosque:.4f}, AUC-ROC: {auc_roc_bosque:.4f}")

El mejor resultado para Random Forest es con los siguientes estimadores:
Estimadores: 100, Profundidad: 15
Accuracy: 0.9498
Precisión: 0.9998
Recall: 0.8999 
F1-score: 0.9472
AUC-ROC: 0.9498

resultados_random_forest = []

# Agregar la combinación específica al DataFrame
resultados_random_forest.append({
    'Modelo': 'Random Forest',
    "Estimadores": 100,
    "Profundidad": 15,
    "Accuracy": 0.9498,
    "Precision": 0.9998,
    "Recall": 0.8999,
    "F1-Score": 0.9472,
    "AUC-ROC": 0.9498
})

# Convertir la lista de resultados en un DataFrame
df_resultados_random_forest = pd.DataFrame(resultados_random_forest)

df_resultados_random_forest

## Test

#Regresion logistica
reg_log.fit(X_prueba_caracteristica, Y_prueba_objetivo)
Y_pred_test = reg_log.predict(X_prueba_caracteristica)

accuracy_rl_test = accuracy_score(Y_prueba_objetivo, Y_pred_test)
precision_rl_test = precision_score(Y_prueba_objetivo, Y_pred_test, zero_division=0)
recall_rl_test = recall_score(Y_prueba_objetivo, Y_pred_test)
f1_rl_test = f1_score(Y_prueba_objetivo, Y_pred_test)
auc_roc_rl_test = roc_auc_score(Y_prueba_objetivo, Y_pred_test)

resultados_lr_test = []
resultados_lr_test.append({
    'Modelo': 'Regresión Logística',
    'Accuracy': accuracy_rl_test,
    'Precisión': precision_rl_test,
    'Recall': recall_rl_test,
    'F1-score': f1_rl_test,
    'AUC-ROC': auc_roc_rl_test
})
df_rl_test = pd.DataFrame(resultados_lr_test)
df_rl_test

#Arbo de decision 
arbol_clf = DecisionTreeClassifier(max_depth=7, min_samples_split=10, criterion="gini", class_weight="balanced", random_state=42)
arbol_clf.fit(X_prueba_caracteristica, Y_prueba_objetivo)
Y_pred_test = arbol_clf.predict(X_prueba_caracteristica)

accuracy_bosque_test = accuracy_score(Y_prueba_objetivo, Y_pred_test)
precision_bosque_test = precision_score(Y_prueba_objetivo, Y_pred_test, zero_division=0)
recall_bosque_test = recall_score(Y_prueba_objetivo, Y_pred_test)
f1_bosque_test = f1_score(Y_prueba_objetivo, Y_pred_test)
auc_roc_bosque_test = roc_auc_score(Y_prueba_objetivo, Y_pred_test)

resultados_decision_test = []

# Agregar la combinación específica al DataFrame
resultados_decision_test.append({
    'Modelo': 'Decision Tree',
    "Profundidad": 14,
    "Split": 10,
    "Accuracy": accuracy_bosque_test,
    "Precision": precision_bosque_test,
    "Recall": recall_bosque_test,
    "F1-Score": f1_bosque_test,
    "AUC-ROC": auc_roc_bosque_test
})

# Convertir la lista de resultados en un DataFrame
df_resultados_decision_tree_test = pd.DataFrame(resultados_decision_test)

df_resultados_decision_tree_test

#Random Forest
bosque_clf = RandomForestClassifier(n_estimators=100, max_depth=15, min_samples_split=5, class_weight="balanced_subsample", random_state=42)
bosque_clf.fit(X_prueba_caracteristica, Y_prueba_objetivo)
Y_pred_test = bosque_clf.predict(X_prueba_caracteristica)

accuracy_bosque_test = accuracy_score(Y_prueba_objetivo, Y_pred_test)
precision_bosque_test = precision_score(Y_prueba_objetivo, Y_pred_test, zero_division=0)
recall_bosque_test = recall_score(Y_prueba_objetivo, Y_pred_test)
f1_bosque_test = f1_score(Y_prueba_objetivo, Y_pred_test)
auc_roc_bosque_test = roc_auc_score(Y_prueba_objetivo, Y_pred_test)

resultados_random_forest_test = []

# Agregar la combinación específica al DataFrame
resultados_random_forest_test.append({
    'Modelo': 'Random Forest',
    "Estimadores": 100,
    "Profundidad": 15,
    "Accuracy": accuracy_bosque_test,
    "Precision": precision_bosque_test,
    "Recall": recall_bosque_test,
    "F1-Score": f1_bosque_test,
    "AUC-ROC": auc_roc_bosque_test
})

# Convertir la lista de resultados en un DataFrame
df_resultados_random_forest_test = pd.DataFrame(resultados_random_forest_test)

df_resultados_random_forest_test
